fix: изменен тип переменной lat и lon класса Event

+++ ewm/src/main/java/ru/practicum/explorewithme/event/dto/LocationDto.java
@@ -12,6 +12,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class LocationDto {
-    private int lat;
-    private int lon;
+    private double lat;
+    private double lon;
 }
